+++ src/lib/SECP256K1Wrapper.nim
@@ -75,7 +75,7 @@ proc `$!`*(pubKeyArg: secp256k1_pubkey): string {.raises: [ValueError].} =
     result = ""
     #Turn the byte array into a hex string.
     for b in bytes:
-        result = result & b.toHex()
+        result = result & uint8(b).toHex()
 
 #Generates a signature from a hex string.
 proc secpSignature(sigArg: string): secp256k1_ecdsa_signature {.raises: [ValueError].} =
@@ -114,7 +114,7 @@ proc `$`(sigArg: secp256k1_ecdsa_signature): string {.raises: [ValueError].} =
     #Turn the byte array into a hex string.
     result = ""
     for b in bytes:
-        result = result & b.toHex()
+        result = result & uint8(b).toHex()
 
 #Sign a message (hash).
 proc secpSign*(privKeyArg: array[32, cuchar], hashArg: string): string {.raises: [ValueError].} =
